pyproject.toml
[project]
name = "webcrawler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.10",
    "selenium>=4.34.2",
    "streamlit>=1.48.1",
    "uvicorn>=0.35.0",
    "webdriver-manager>=4.0.2",
]
